third_party/WebKit/LayoutTests/typedcssom/stylevalue-objects/parse-relative-url.html

<!doctype html>
<meta charset="utf-8">
<title>parsing relative url tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#dom-cssstylevalue-parse">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/testhelper.js"></script>
<body>
<script>
'use strict';

const gTestUrl = '../resources/1x1-green.png';

function loadImageResource(imageValue) {
  // Set a CSSURLImageValue on an element so it can be loaded.
  let target = newDivWithStyle('');
  document.body.append(target);
  target.attributeStyleMap.set('background-image', imageValue);

  // add a new Image element to know if the image resource has been loaded
  let image = new Image();
  image.src = imageValue.url;
  return image;
}

async_test(t => {
  const result = CSSStyleValue.parse('background-image', 'url('+gTestUrl+')');
  let image = loadImageResource(result);

  image.addEventListener('load', t.step_func_done(() => {
    assert_equals(result.state, 'loaded');
    assert_equals(result.intrinsicWidth, 1);
    assert_equals(result.intrinsicHeight, 1);
    assert_equals(result.intrinsicRatio, 1);
    assert_equals(result.url, gTestUrl);
  }));
}, 'Loading a CSSURLImageValue which parsed using CSSStyleValue.parse from a URL sets its state to loaded');

</script>
